installer.sh revised

+++ installer.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
+
 FLENAME=~/dveltool_installer.sh
-rm $FLENAME
+
+if [ -f "$FLENAME" ]; then
+	rm $FLENAME
+fi
+
 wget -O $FLENAME https://raw.githubusercontent.com/dveltool/toolchain/master/installer/dveltool_installer.sh
 chmod +x $FLENAME
 sh $FLENAME
